+++ dejafu/CHANGELOG.markdown
@@ -7,8 +7,12 @@ This project is versioned according to the [Package Versioning Policy](https://p
 *de facto* standard Haskell versioning scheme.
 
 
-unreleased
-----------
+0.7.2.0
+-------
+
+- **Date**    2017-09-16
+- **Git tag** [dejafu-0.7.2.0][]
+- **Hackage** https://hackage.haskell.org/package/dejafu-0.7.2.0
 
 ### Test.DejaFu.STM
 
@@ -23,6 +27,8 @@ unreleased
     This makes it consistent with the `Ord` instances, and is also a small performance gain.
 
 - Now compatible with concurrency-1.2.0.0.
+
+[dejafu-0.7.2.0]: https://github.com/barrucadu/dejafu/releases/tag/dejafu-0.7.2.0
 
 
 ---------------------------------------------------------------------------------------------------